// src/app/models/rank.model.ts
export enum Rank {

  White = -10,
  Yellow = -9,
  Orange = -8,
  Blue = -7,
  Green2 = -6,
  Green1 = -5,
  Brown2 = -4,
  Brown1 = -3,
  Red = -2,
  RedBlack = -1,
  Black1 = 1,
  Black2= 2,
  Black3 = 3,
  Black4 = 4,
  Black5 = 5,
  Black6 = 6,
  Black7 = 7,
  Black8 = 8,
  Black9 = 9,
  Black10 = 10

}